# feat: ProtobufParse.formatOpenMetricsFloat: improve float formatting …