@beorn7 2024-10-16
I have quickly dived into the `fmt` code. It already uses a `sync.Pool` for it's working buffers, so I would think the returns from this PR are diminishing the more this code path is used, in particular in use cases where many parsers are instantiated.

However, with my idea in the comments above (i.e. also append `.0` to the buffer and only then do the string conversion), we have a very substantial gain for the cases where a `.0` actually has to be appended (which is plausible to happen a lot, e.g. with a histogram with many buckets with integer boundaries