@machine424 2024-10-17
> I have quickly dived into the fmt code. It already uses a sync.Pool for it's working buffers, so I would think the returns from this PR are diminishing the more this code path is used, in particular in use cases where many parsers are instantiated (in which case the change here could even be a net negative).

`BenchmarkParse` does instantiate a new parser on each iteration (imitating what happens in `scrape.append()`), and still in `BenchmarkParse` use case, having a per parser buffer looks more optimal that fmt's (I didn't into its pooling code).

I had tried a global sync.Pool, seen no difference, but I wasn't aware a new parser was created on every scrape/iteration. Avoiding a requisite 24bytes per scrape will always be better, I'll go back to sync.Pool.