@machine424 2024-10-18
>How do the benchmark look like with the current state?

I added it to the description.

>Let's maybe add the case where .0 has to be appended to the benchmarks to see how well it works?
Saw this coming, I realized I was asking for trouble when I pointed out that the tests didn’t cover that case :) Added one.

Comparing it with the "alter in string" version
```
                                                   │   v0.txt    │            v1.txt             │
                                                   │   sec/op    │   sec/op     vs base          │
Parse/data=createTestProtoBuf()/parser=promproto-2   33.85µ ± 4%   34.39µ ± 3%  ~ (p=0.529 n=10)

                                                   │    v0.txt    │             v1.txt             │
                                                   │     B/s      │     B/s       vs base          │
Parse/data=createTestProtoBuf()/parser=promproto-2   108.3Mi ± 4%   106.6Mi ± 3%  ~ (p=0.529 n=10)

                                                   │    v0.txt    │               v1.txt                │
                                                   │     B/op     │     B/op      vs base               │
Parse/data=createTestProtoBuf()/parser=promproto-2   40.15Ki ± 0%   40.14Ki ± 0%  -0.04% (p=0.000 n=10)

                                                   │   v0.txt   │              v1.txt               │
                                                   │ allocs/op  │ allocs/op   vs base               │
Parse/data=createTestProtoBuf()/parser=promproto-2   876.0 ± 0%   870.0 ± 0%  -0.68% (p=0.000 n=10)
```